+++ scripts/martech.js
@@ -1,0 +1,82 @@
+/*
+ * Copyright 2021 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+import {
+  loadScript
+} from '/scripts/common.js';
+
+/**
+ * sets marketing tech context
+ */
+
+function setMarTechContext() {
+  var env='dev';
+  var hostname=window.location.hostname;
+  if (hostname.includes('staging')) env='stage';
+  if (hostname == 'blog.adobe.com') env='production';
+
+  var isDX=false;
+  document.querySelectorAll('main>div:last-of-type>p').forEach(($p) => {
+    if ($p.innerHTML.includes('Products:') || $p.innerHTML.includes('Topics:')) {
+      if (checkDX($p.innerHTML.split(':')[1])) {
+        isDX=true;
+      }
+    }
+  });
+
+  var accounts='';
+  if (isDX) {
+    if (env == 'production') {
+      accounts='adbadobedxprod';
+    }
+    if (env == 'stage') {
+      accounts='adbadobedxqa';
+    }
+  }
+
+  window.marketingtech = {
+    adobe: {
+      launch: {
+        property: 'global',
+        environment: env  // “production” for prod/live site or “stage” for qa/staging site
+      },
+      analytics: {
+        additionalAccounts: accounts // additional report suites to send data to “,” separated  Ex: 'RS1,RS2'
+      },
+      target: true,    // if target needs to be enabled else false
+      audienceManager: true    // if audience manager needs to be enabled else false
+    }
+  };
+  // console.log(window.marketingtech)
+}
+
+/**
+ * sets digital data
+ */
+
+function setDigitalData() {
+  var langMap={'en': 'en-US'};
+  var lang=window.blog.language;
+  if (langMap[lang]) lang=langMap[lang];
+  digitalData._set('page.pageInfo.language', lang);
+  console.log(lang);
+}
+
+setMarTechContext();
+
+loadScript('https://www.adobe.com/marketingtech/main.min.js', () => {
+  setDigitalData();
+});
+
+loadScript('https://www.adobe.com/etc.clientlibs/globalnav/clientlibs/base/feds.js').id = 'feds-script';
+loadScript('https://static.adobelogin.com/imslib/imslib.min.js');
+
